--- src/components/Apply3.js
@@ -1,18 +1,0 @@
-import React from 'react'
-
-function Apply3() {
-	return (
-		<div className="main">
-      <div className="leftBox">
-        <img alt="" src="/image/logo-white 1.png" className="logo" />
-        <img alt="" src="/image/remoteStack.png" className="remoteStack" />
-      </div>
-      <div className="rightBox">
-        <h2 className="title">JOIN AS A GLOBAL DEVELOPER</h2>
-        
-      </div>
-    </div>
-	)
-}
-
-export default Apply3
